Bump test build to v07072024
-Added Experimental Network Adapter Link Local IPv4 Address detection support

The way it works is that a separate class, with the help of JAVA's NetworkInterface and InetAddress classes maps all the active adapters and their IPv4 address in a map.

FeL also returns its own map.

These two maps are then compared to find a common key. Luckily, FeL's "Name" key provides a value as the key we use for our new map, which helps us link the IPv4 address.

GOT DAM THIS IS EVEN MORE COMPLICATED TO EXPLAIN I HOPE I REMEMBER HOW THIS WORKED WHEN I LOOK BACK. FOR THE LOVE OF GOD PLS UPGRADE TO WINDOWS 10/11 ALREADY